Version 2.5.0
Added more outputformats (+hash, +json) for the statistic stuff.
Personally I do not have use for this, but it was a nice, little task,
doesn't seem to hurt anything and just seems to be the right thing to
do.
Since this is a new functionality, the minor version number just
increased to 2.5
Updated also the Changelog and the help for the stat method.